pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "browsy-automation"
version = "0.1.0"
description = "Intelligent web automation powered by Playwright, MCP, and OpenAI LLM"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Browsy Team", email = "hello@browsy.dev"},
]
keywords = [
    "automation",
    "playwright",
    "mcp",
    "llm",
    "openai",
    "web-scraping",
    "browser-automation",
    "ai-agent",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "mcp-agent>=0.0.14",
    "openai>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "aiosqlite>=0.19.0",
]
cli = [
    "typer>=0.9.0",
    "rich>=13.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "browsy-automation[api,cli]",
]

[project.urls]
Homepage = "https://github.com/muneebanwar69/browsy-automation"
Documentation = "https://github.com/muneebanwar69/browsy-automation#readme"
Repository = "https://github.com/muneebanwar69/browsy-automation"
Issues = "https://github.com/muneebanwar69/browsy-automation/issues"

[project.scripts]
browsy = "browsy.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/browsy"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.black]
line-length = 120
target-version = ["py39"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
